update function for production
